xfs: prevent a WARN_ONCE() in xfs_ioc_attr_list()
The "bufsize" comes from the root user.  If "bufsize" is negative then,
because of type promotion, neither of the validation checks at the start
of the function are able to catch it:

	if (bufsize < sizeof(struct xfs_attrlist) ||
	    bufsize > XFS_XATTR_LIST_MAX)
		return -EINVAL;

This means "bufsize" will trigger (WARN_ON_ONCE(size > INT_MAX)) in
kvmalloc_node().  Fix this by changing the type from int to size_t.


(cherry picked from commit 6ed6356b07714e0198be3bc3ecccc8b40a212de4)
Orabug: 34085023

Conflicts:
	fs/xfs/xfs_ioctl.c - The code to be modified is in
	                     xfs_attrlist_by_handle() function rather than in
			     xfs_ioc_attr_list() as mentioned in the upstream
			     commit. Typecast al_hreq.buflen to size_t to
			     cause the same effect as the upstream commit.
	fs/xfs/xfs_ioctl.h - No changes here since there is no change in the
	                     prototype of xfs_ioc_attr_list()
(cherry picked from commit 2431914e54affded86391f892eaeda699bcab608)

+++ fs/xfs/xfs_ioctl.c
@@ -383,7 +383,7 @@ xfs_attrlist_by_handle(
 	if (copy_from_user(&al_hreq, arg, sizeof(xfs_fsop_attrlist_handlereq_t)))
 		return -EFAULT;
 	if (al_hreq.buflen < sizeof(struct attrlist) ||
-	    al_hreq.buflen > XFS_XATTR_LIST_MAX)
+	    (size_t)(al_hreq.buflen) > XFS_XATTR_LIST_MAX)
 		return -EINVAL;
 
 	/*
